refactor: move subroute enum to enums folder
