+++ frontend/dashboard/context/HeaderContext/index.ts
@@ -1,6 +1,1 @@
-export {
-  HeaderContext,
-  type HeaderContextType,
-  SelectedContextType,
-  Subroute,
-} from './HeaderContext';
+export { HeaderContext, type HeaderContextType, SelectedContextType } from './HeaderContext';
