+++ frontend/dashboard/enums/Subroute.ts
@@ -1,0 +1,6 @@
+import { APP_DASHBOARD_BASENAME, ORG_LIBRARY_BASENAME } from 'app-shared/constants';
+
+export enum Subroute {
+  AppDashboard = APP_DASHBOARD_BASENAME,
+  OrgLibrary = ORG_LIBRARY_BASENAME,
+}
