+++ frontend/dashboard/pages/PageLayout/DashboardHeader/DashboardHeader.test.tsx
@@ -2,7 +2,8 @@ import React from 'react';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { DashboardHeader } from './DashboardHeader';
-import { HeaderContext, SelectedContextType, Subroute } from 'dashboard/context/HeaderContext';
+import { HeaderContext, SelectedContextType } from 'dashboard/context/HeaderContext';
+import { Subroute } from 'dashboard/enums/Subroute';
 import { useParams } from 'react-router-dom';
 import { textMock } from '@studio/testing/mocks/i18nMock';
 import { type User } from 'app-shared/types/Repository';
